+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.29.0] - 2018-11-06
+### Added
+- Added function to bulk add stat subscriptions
+- Added function to bulk remove stat subscriptions
+
 ## [8.28.1] - 2018-11-06
 ### Changed
 - Updating functions for Access Control to match fixes in APIs.